worker: move to its own package
Package `common` is an antipattern.
